+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 - Command line option --severity/-S for filtering by minimum severity
 - SC2236/SC2237: Suggest -n/-z instead of ! -z/-n
+- SC2238: Warn when redirecting to a known command name, e.g. ls > rm
 ### Changed
 - Most warnings now have useful end positions
 - SC1117 about unknown double-quoted escape sequences has been retired